Extract orchestrator step execution so orchestrator.py stays under 1k lines.
